VfioUserServer: remove exception for PassthroughDevice
